+++ lib/src/clixon_yang_parse.h
@@ -51,7 +51,8 @@ struct ys_stack{
 };
 
 struct clicon_yang_yacc_arg{ /* XXX: mostly unrelevant */
-    char                 *yy_name;         /* Name of syntax (for error string) */
+    char                 *yy_name;         /* Name of syntax, typically filename 
+					      (for error string) */
     int                   yy_linenum;      /* Number of \n in parsed buffer */
     char                 *yy_parse_string; /* original (copy of) parse string */
     void                 *yy_lexbuf;       /* internal parse buffer from lex */